requirements.txt
anthropic
openai
pydantic
pydantic-settings
python-dotenv
rich
click
httpx
pyautogui
pywin32

